test refactor

+++ src/main/java/org/juraj/durej/app/database/UserRepository.java
@@ -1,6 +1,7 @@
 package org.juraj.durej.app.database;
 
 import java.util.List;
+import java.util.concurrent.TimeUnit;
 import java.util.concurrent.locks.Lock;
 import java.util.concurrent.locks.ReentrantLock;
 import org.hibernate.Session;
@@ -12,16 +13,23 @@ import org.slf4j.LoggerFactory;
 public class UserRepository {
 
   private static final Logger log = LoggerFactory.getLogger(UserRepository.class);
+  private static final int TRY_LOCK_MAX_TIMES = 10;
   Transaction transaction = null;
-  private Lock lock = new ReentrantLock();
+  private final Lock lock = new ReentrantLock();
 
   public void addUser(User user) {
     try (Session session = HibernateUtils.getSessionFactory().openSession()) {
-      lock.lock();
-      transaction = session.beginTransaction();
-      session.persist(user);
-      transaction.commit();
 
+      for (int i = 0; i < TRY_LOCK_MAX_TIMES; i++) {
+        if (lock.tryLock(100, TimeUnit.MILLISECONDS)) {
+
+          transaction = session.beginTransaction();
+          session.persist(user);
+          transaction.commit();
+
+          break;
+        }
+      }
     } catch (Exception e) {
       if (transaction != null) {
         transaction.rollback();
@@ -38,9 +46,6 @@ public class UserRepository {
       return session.createQuery("SELECT u FROM User u", User.class).list();
 
     } catch (Exception e) {
-      if (transaction != null) {
-        transaction.rollback();
-      }
       log.error(e.getMessage());
     }
     return List.of();
@@ -48,12 +53,17 @@ public class UserRepository {
 
   public void deleteAllUsers() {
     try (Session session = HibernateUtils.getSessionFactory().openSession()) {
-      transaction = session.beginTransaction();
-      lock.lock();
-      session.createQuery("DELETE FROM User").executeUpdate();
 
-      transaction.commit();
+      for (int i = 0; i < TRY_LOCK_MAX_TIMES; i++) {
+        if (lock.tryLock(100, TimeUnit.MILLISECONDS)) {
 
+          transaction = session.beginTransaction();
+          session.createQuery("DELETE FROM User").executeUpdate();
+          transaction.commit();
+
+          break;
+        }
+      }
     } catch (Exception e) {
       if (transaction != null) {
         transaction.rollback();
